[update] Philosopher’s Walk(14956) - 틀렸습니다 해결

+++ baekjoon/baaarkingdog/0x0B/Philosopher’s Walk(14956).py
@@ -4,8 +4,8 @@ n, m = map(int, input().split())
 def s(n, m):
     if n == 2:
         return (1, 1) if m == 0 else (1, 2) if m == 1 else (2, 2) if m == 2 else (2, 1)
-    k = n/2
-    p = m / (k*k)
+    k = int(n/2)
+    p = int(m / (k*k))
     a, b = s(n >> 1, m % (k*k))
     return (b, a) if p == 0 else (a, b+k) if p == 1 else (a+k, b+k) if p == 2 else (k-b+1+k, k-a+1)
